+++ CHANGELOG.md
@@ -6,9 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.1.0] - 2018-05-20
+### Added
+- New class for PMT calculation (FinancialCalculator::Pmt)
+
+### Removed
+
+### Changed
+
 ## [3.0.0] - 2018-05-12
 ### Added
-- New class for Present Value calculation (FinancialCalculator::PV)
+- New class for Present Value calculation (FinancialCalculator::Pv)
 
 ### Removed
 - Monkey patched cashflow methods on Array